test(api): move api spec test to dedicated module
